requirements.txt
asgiref==3.2.10
astroid==2.4.2
dj-database-url==0.5.0
Django==3.1.1
django-appconf==1.0.4
django-ckeditor==6.0.0
django-imagekit==4.0.2
django-js-asset==1.2.2
gunicorn==20.0.4
isort==5.5.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
pep8==1.7.1
pilkit==2.0
Pillow==7.2.0
psycopg2==2.8.6
pylint==2.6.0
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.2.0
wrapt==1.12.1
